pyproject.toml
[tool.poetry]
name = "workshop-airflow"
version = "0.1.0"
description = ""
authors = ["Luiz Fernando <luizfsoliveira.lm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^1.4.36"
apache-airflow = "^2.10.4"
pydantic = "^2.10.5"
psycopg2-binary = "^2.9.10"
apache-airflow-providers-postgres = "^6.0.0"
astro-sdk-python = {extras = ["duckdb"], version = "^1.8.1"}
airflow-provider-duckdb = "^0.2.0"
duckdb = "^1.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
